style: move sidebar-sidecar-layout into partials folder
